server: split operator from balancer file.
